Started implementing FileTests and File
-Refactored code to use more FieldInfo structs and constants
-Updated LibCppBinary and googletest to latest versions
-Wrote first file test
-All file tests pass

+++ src/Spc/Id666/Extended/Data.cpp
@@ -23,7 +23,7 @@ Binary::ChunkHeader Data::Header() const
 {
     Binary::ChunkHeader header;
     header.id.SetValue("xid6");
-    header.dataSize.SetValue(Size());
+    header.dataSize.SetValue(static_cast<uint32_t>(Size()));
     return header;
 }
